fix warnings

+++ targets/RT/USER/lte-ue.c
@@ -2286,7 +2286,7 @@ int init_timer_thread(void)
 /* HACK: this function is needed to compile the UE
  * fix it somehow
  */
-int8_t find_dlsch(uint16_t rnti,
+int find_dlsch(uint16_t rnti,
                   PHY_VARS_eNB *eNB,
                   find_type_t type)
 {
